Implement OpenSSL's EVP Digest calls for s2n_hash
Migrate the s2n_hash implementations to the EVP Digest APIs in
OpenSSL. These API's are the recommended hash digest
implementations and are the required APIs for use in FIPS mode
when linked with a FIPS-capable libcrypto.

The changes to these implementations are transparent to all
existing tests and uses of the s2n_hash functions.